Add combined plural and sprintf function
And exposing it to the public api. The function uses the first argument
to get the correct plural translation, the second argument is the
string to translate and all the following arguments are used for
sprintf. The function also replaces __count__ and %d with %s to use in
the sprintf function.