Exposing the translate and plural function to helper
The original t and plural functions are exposed to use directly in eg.
Express routes.

They are exposed as ‘__’ and ‘_p’, and the ’t’ function is kept as is.
The function exposed as ’t’ does not take several arguments, and thus
have problem with a sprintf string with multiple variables. The new
exposed function may have problem with dynamic language control, and
has only been tested with a fixed language for the Node.js application.