Default setJqueryExt to false in node environment
Having this option defaulted to true causes an error on init if the jquery npm module is installed since jQuery exports a function that throws in a pure node environment - attempting to set $.fn.i18n fails since $.fn is undefined.

Adding jquery to dev dependencies as a way of failing test for this bug.