Fix lookup with `fallbackToDefaultNS` option
Running a lookup with fallbackToDefaultNS flag was not actually doing the fallback, it was simply running the same lookup again with all of the same options - in particular the same namespace option - so it never found the fallback value.

When running the fallback lookup set the ns option to be the default namespace and re-run the lookup.